added some styling to topic compoenent-- need some more just doing this push so that I can get the newest pull version
