+++ client/src/pages/LandingPage.jsx
@@ -1,34 +1,34 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
+import React from "react";
+import { Link } from "react-router-dom";
 
 const LandingPage = () => {
-
-
   return (
     <main>
-      <div className='container'>
-        <div className='row'>
-          <div className='col-12 col-md-6'>
-          <div className="card">
-                <img src="https://via.placeholder.com/300" className="card-img-top" alt="..." />
-              </div>
+      <div className="container">
+        <div className="row">
+          <div className="col-12 col-md-6">
+            <div className="card">
+              <img
+                src="https://via.placeholder.com/300"
+                className="card-img-top"
+                alt="..."
+              />
+            </div>
           </div>
 
-          <div className='col-12 col-md-6'>
+          <div className="col-12 col-md-6">
             <div className="card">
-              <img src="https://via.placeholder.com/300" className="card-img-top" alt="..." />
+              <img
+                src="https://via.placeholder.com/300"
+                className="card-img-top"
+                alt="..."
+              />
               <div className="card-body">
-                <Link
-                  to="/topic"
-                  className='btn btn-primary'
-                >
-                Take a Quiz
+                <Link to="/topics" className="btn btn-primary">
+                  Take a Quiz
                 </Link>
-                <Link
-                  to="/login"
-                  className='btn btn-primary'
-                >
-                Login
+                <Link to="/login" className="btn btn-primary">
+                  Login
                 </Link>
               </div>
             </div>
@@ -36,8 +36,7 @@ const LandingPage = () => {
         </div>
       </div>
     </main>
-
-  )
-}
+  );
+};
 
 export default LandingPage;
