+++ client/src/pages/Topic/index.css
@@ -1,0 +1,11 @@
+a:link { 
+  text-decoration: none; 
+}
+
+.topicsBtn:hover { 
+  background-color: green; 
+}
+
+a:hover {
+  color: white;
+}
